requirements.txt
aiohttp==3.9.3
asyncio==3.4.3
python-socketio==5.8.0
msgspec==0.18.6
